+++ src/commands/enlargeemoji.js
@@ -3,7 +3,7 @@ const { SlashCommandBuilder } = require('@discordjs/builders');
 module.exports = {
 	data: new SlashCommandBuilder()
 		.setName('enlargeemoji')
-		.setDescription('Pastes the emoji\'s url to chat.')
+		.setDescription('Pastes a custom emoji\'s url to chat.')
 		.addStringOption(option =>
 			option.setName('emoji')
 				.setDescription('The emoji to display.')
@@ -13,17 +13,26 @@ module.exports = {
 		// Validates emoji option.
 		const stringEmoji = interaction.options.getString('emoji');
 		const re = /<?(a)?:?(\w{2,32}):(\d{17,19})>?/;
-		let emoji = stringEmoji.match(re);
-		try {
-			emoji = await interaction.guild.emojis.fetch(emoji[3]);
-		}
-		catch {
-			return interaction.reply({ content: 'No custom emoji found in string.', ephemeral: true });
-		}
+		const regexEmoji = stringEmoji.match(re);
+		if (regexEmoji) {
 
-		// Sends emoji to chat.
-		if (emoji) {
-			return interaction.reply({ content: `${emoji.url}` });
+			// Checks for if emoji is animated or not.
+			let url;
+			if (regexEmoji[1]) {
+				url = `https://cdn.discordapp.com/emojis/${regexEmoji[3]}.gif`;
+			}
+			else {
+				url = `https://cdn.discordapp.com/emojis/${regexEmoji[3]}.png`;
+			}
+
+			return interaction.reply({ content: `${url}` });
+		}
+		else {
+			return interaction.reply({
+				content: 'No custom emoji was found in your message!',
+				ephemeral: true,
+			});
+
 		}
 
 	},